Fitting the data

The reconstruction and the fitting parts are totally independent.

The fitting functions expects a multidimensionnal-array with the echoes along the last dimension. The images should not be in complex, remember to use abs.(img)

Exponential $T_2$ fitting + noise

In the publication we used an analytical model to fit the exponential decay of the echoes :

\[S(t) = \sqrt{\left(M_0 \ exp(-\frac{t}{T_2})\right)^2 + 2 \ L \ \sigma_g^2}\]

where L is the number of coil and $\sigma_g$ corresponds to the gaussian noise level on image

You can perform this fit with the following function :

Subspace_MESE.T2Fit_exp_noise — Function

    T2Fit_exp_noise(ima::Array{T,N}, t::AbstractVector{T}; removePoint::Bool=true, L::Int=1, mask = nothing) where {T<:Real,N}

Fit the relaxation parameters T2 with the equation : $S(t) = \sqrt{(M_0 \exp(-\frac{t}{T2}))^2 + 2 L \sigma_g^2}$ where L est le nombre de canaux, et $\sigma_g$ le bruit gaussien sur les image

Arguments

ima::Array{T,N}: multi-dimensionnal images. Last dimension stores the temporal dimension

t::AbstractVector{<:Real}: times vector in ms

p0=nothing: starting values for fit, if empty p0=[maximum(ima),30,maximum(ima)*0.1]

Keywords

removePoint::Bool=true: remove the first point before fitting

L::Int=1: Number of coil elements

mask::

Returns

fit_params : parameter maps last dimension stores the following maps (M₀ , T₂ , σ)

Bibliography

Cárdenas-Blanco A, Tejos C, Irarrazaval P, Cameron I. Noise in magnitude magnetic

resonance images. Concepts Magn Reson Part A [Internet]. 2008 Nov;32A(6):409?16. Available from: http://doi.wiley.com/10.1002/cmr.a.20124

Feng Y, He T, Gatehouse PD, Li X, Harith Alam M, Pennell DJ, et al. Improved MRI R 2 *

relaxometry of iron-loaded liver with noise correction. Magn Reson Med [Internet]. 2013 Dec;70(6):1765?74. Available from: http://doi.wiley.com/10.1002/mrm.24607

The keyword removePoint can be used to delete the first point in the TE vector as well as the first temporal volume in order to reduce the sensitivity of the fit to the stimulated echo.

EPG $T_2$ fitting + noise (WIP)

Another possibility is to fit the equation with an EPG model that, in addition to $M_0$/$T_2$/$\sigma$, also fit the $B_1$ field.

With the current implementation, the fit is not robust enough and also takes too long to use on a volumic volume.

T2Fit_epg_noise